Improved AggregateOperator

+++ odysseus4iot_eclipse-project/src/percom/graph/operator/meta/Operator.java
@@ -5,10 +5,10 @@ import percom.graph.Vertex;
 public class Operator extends Vertex
 {
 	public Schema inputSchema = null;
-	public Integer inputRate = null;
+	public Double inputRate = null;
 	public String inputName = null;
 	
 	public Schema outputSchema = null;
-	public Integer outputRate = null;
+	public Double outputRate = null;
 	public String outputName = null;
 }
